+++ pages/page.py
@@ -3,13 +3,13 @@ from urllib import parse
 
 class Page:
     BASE_URL = 'https://ok.ru/'
-    PATH = ''
 
-    def __init__(self, driver):
+    def __init__(self, driver, path=''):
         self.driver = driver
+        self.path = path
 
     def open(self):
-        url = parse.urljoin(self.BASE_URL, self.PATH)
+        url = parse.urljoin(self.BASE_URL, self.path)
         self.driver.get(url)
         self.driver.maximize_window()
 
@@ -18,3 +18,5 @@ class Component:
 
     def __init__(self, driver):
         self.driver = driver
+
+
